Separate CanvasView + MediaPlayer from Activity into CanvasFragment
Instead of doing everything in the activity, the canvas content is now part of a fragment,
which gets added when a valid host has been configured.
The action bar buttons are processed by the activity and the corresponding preferences changed.
The fragment then reacts to those changes and updates its content accordingly.