Replace RTSP server with MPEG-DASH server
ExoPlayer does not support RTSP yet, so only DASH and HLS are an option.
With this configuration of ffmpeg, the latency can be brought down to 2 seconds.
While RTSP can provide a much better latency, it requires a small buffer size
to do so and this is only possible with ExoPlayer, but not Android's MediaPlayer.
When ExoPlayer supports RTSP, this should improve latency even more.